+++ src/main/scala/apdl/generation/CLikeTransformCodeGenerator.scala
@@ -1,0 +1,120 @@
+package apdl.generation
+
+import apdl.ApdlCodeGenerationException
+import apdl.parser._
+
+class CLikeTransformCodeGenerator {
+  def apply(apdlAst: ApdlAst): String = apdlAst match {
+    case e: Expr => e match {
+      case Add(left, right) => s"(${apply(left)} + ${apply(right)})"
+      case Mul(left, right) => s"(${apply(left)} * ${apply(right)})"
+      case Sub(left, right) => s"(${apply(left)} - ${apply(right)})"
+      case Div(left, right) => s"(${apply(left)} / ${apply(right)})"
+      case Cast(tfTyp, expr) => s"((${apply(tfTyp)})${apply(expr)})"
+      case Literal(value) => s"$value"
+      case Symbol(name) => s"$name"
+      case FunctionCall(funcName, args) => s"$funcName(${args map apply mkString ","})"
+      case ArrayAccess(array, field) => s"${apply(array)}[${apply(field)}]"
+      case VarAssignement(target, value) => s"${apply(target)} = ${apply(value)}"
+      case True() => s"1"
+      case False() => s"0"
+      case Or(left, right) => s"(${apply(left)} || ${apply(right)})"
+      case And(left, right) => s"(${apply(left)} && ${apply(right)})"
+      case Not(booleanExpr) => s"(! ${apply(booleanExpr)})"
+      case Greater(left, right) => s"(${apply(left)} > ${apply(right)})"
+      case Smaller(left, right) => s"(${apply(left)} < ${apply(right)})"
+      case GreaterEquals(left, right) => s"(${apply(left)} >= ${apply(right)})"
+      case SmallerEquals(left, right) => s"(${apply(left)} <= ${apply(right)})"
+      case Equals(left, right) => s"(${apply(left)} == ${apply(right)})"
+      case NotEquals(left, right) => s"(${apply(left)} != ${apply(right)})"
+    }
+    case t: TfRetTyp => t match {
+      case TfInt => "int"
+      case TfLong => "long"
+      case TfByte => "byte"
+      case TfShort => "short"
+      case TfChar => "char"
+      case TfDouble => "double"
+      case TfFloat => "float"
+      case TfVoid => "void"
+      case TfBoolean => "bool"
+      case TfArray(typ: TfTyp) => s"${apply(typ)}*"
+    }
+    case TypedIdentifier(name, typ) => s"${apply(typ)} $name"
+    case s: Statement => s match {
+      case While(cond, statement) =>
+        s"""
+           |while(${apply(cond)})
+           |  ${apply(statement)}
+         """.stripMargin
+      case DoWhile(cond, statement) =>
+        s"""
+           |do
+           |  ${apply(statement)}
+           |while (${apply(cond)});
+         """.stripMargin
+      case IfThenElse(cond, trueBranch, falseBranch) =>
+        s"""
+           |if(${apply(cond)})
+           |  ${apply(trueBranch)}
+           |else
+           |  ${apply(falseBranch)}
+         """.stripMargin
+      case IfThen(cond, ifTrue) =>
+        s"""
+           |if(${apply(cond)})
+           |  ${apply(ifTrue)}
+         """.stripMargin
+      case Return(expr) => s"return ${apply(expr)};"
+      case Break() => s"break;"
+      case Continue() => s"continue;"
+      case Block(statements) =>
+        s"""|{
+            |  ${statements map apply mkString "\n"}
+            |}""".stripMargin
+      case ExpressionStatement(expression) => s"${apply(expression)};"
+      case decl: Declaration => decl match {
+        // No { or } because the body is always a block and the block already generate the {}
+        case FunctionDecl(header, body) =>
+          s"""
+             |${apply(header.resultType)} ${header.identifier} (${header.parameters map apply mkString ","})
+             |  ${apply(body.body)}
+           """.stripMargin
+        case NewVal(symbol, typ, init) => s"${apply(typ)} ${apply(symbol)} = ${apply(init)};"
+        case NewVar(symbol, typ, init) => init match {
+          case Some(value) => s"${apply(typ)} ${apply(symbol)} = ${apply(value)};"
+          case None => s"${apply(typ)} ${apply(symbol)};"
+        }
+        case NewArray(symbol, typ, init) => s"${apply(typ)} ${apply(symbol)} = ${apply(init)};"
+      }
+    }
+    case arrayInit: ArrayInit => arrayInit match {
+      case ArrayInitValue(values) => s"{${values map apply mkString ","}"
+      case ArrayInitCapacity(capacity) => s"[${apply(capacity)}]"
+    }
+  }
+
+  def apply(typ: ApdlType): String = typ match {
+    case ApdlType.Int => "int"
+    case ApdlType.Float => "float"
+    case ApdlType.Long => "long"
+    case ApdlType.Bool => "bool"
+    case ApdlType.Double => "double"
+    case ApdlType.Short => "short"
+    case ApdlType.Char => "char"
+    case ApdlType.Byte => "byte"
+    case _ => throw new ApdlCodeGenerationException(s"Unsupported ApdlType : $typ")
+  }
+
+  def strTypeFormater(typ: ApdlType): String = typ match {
+    case ApdlType.Int => "%d"
+    case ApdlType.Float => "%f"
+    case ApdlType.Long => "%l"
+    case ApdlType.Bool => "%d"
+    case ApdlType.Double => "%f"
+    case ApdlType.Short => "%d"
+    case ApdlType.Char => "%c"
+    case ApdlType.Byte => "%c"
+    case _ => throw new ApdlCodeGenerationException(s"Unsupported ApdlType : $typ")
+  }
+}
